Testing/student three backend

+++ Eventy/src/main/java/org/example/eventy/events/services/BudgetItemService.java
@@ -9,6 +9,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.time.LocalDateTime;
 import java.util.ArrayList;
 
 @Service
@@ -21,12 +22,17 @@ public class BudgetItemService {
         budgetItem.setCategory(category);
         budgetItem.setPlannedFunds(allocatedFunds);
         budgetItem.setReservedItems(new ArrayList<>());
-        return budgetItemRepository.save(budgetItem);
+        try {
+            return budgetItemRepository.save(budgetItem);
+        } catch (Exception e) {
+            return null;
+        }
     }
 
     public BudgetItem updateAllocatedFunds(Long id, Double allocatedFunds) {
         BudgetItem budgetItem = budgetItemRepository.findById(id).orElse(null);
         if (budgetItem == null) { return null; }
+        if (allocatedFunds < 0) { return null; }
         budgetItem.setPlannedFunds(allocatedFunds);
         return budgetItemRepository.save(budgetItem);
     }
@@ -36,18 +42,22 @@ public class BudgetItemService {
         return budgetItemRepository.save(budgetItem);
     }
 
-    public boolean deleteBudgetItemSolution(Long id, Long solutionHistoryId) {
+    public boolean deleteBudgetItemSolution(Long id, Long solutionHistoryId, LocalDateTime eventDate) {
         BudgetItem budgetItem = budgetItemRepository.findById(id).orElse(null);
         if (budgetItem == null) { return false; }
         int lengthBefore = budgetItem.getReservedItems().size();
-        budgetItem.getReservedItems().removeIf(v -> v.getId() == solutionHistoryId);
+        budgetItem.getReservedItems().removeIf(v -> (v.getId() == solutionHistoryId && (v.getCancellationDeadline() == null || eventDate.plusDays(v.getCancellationDeadline()).isBefore(LocalDateTime.now()))));
         budgetItemRepository.save(budgetItem);
         int lengthAfter = budgetItem.getReservedItems().size();
         return lengthBefore != lengthAfter;
     }
 
     @Transactional
-    public void deleteBudgetItem(Long id) {
+    public boolean deleteBudgetItem(Long id) {
+        BudgetItem budgetItem = budgetItemRepository.findById(id).orElse(null);
+        if (budgetItem == null) { return false; }
+        if (!budgetItem.getReservedItems().isEmpty()) { return false; }
         budgetItemRepository.deleteById(id);
+        return true;
     }
 }
